Move curl encode functionality back from a separate script
On second thought, that was a silly idea. It also turns out that I never
removed the -u option from the getopts line, so it's back (with the new
and improved implementation) on that option, but with a slight change:
instead of the option taking a single arg, we encode and join all the
non-option arguments. This makes it actually useful for curlicue-setup.